Add hard assumption that type and url are the same
This simplifies the contructor of the base `Resource` to use the `type`
as the `url` as well instead of having to specify separate values for
both.  This is acceptable because according to the JSON API 1.0 spec,
the values for `type` should be plural which usually matches what you
want for the `url` as well.